Updated baseline profiles
