+++ app/build.gradle
@@ -33,6 +33,8 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
         benchmark {
+            initWith buildTypes.release
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro', 'proguard-benchmark.pro'
             signingConfig signingConfigs.debug
             matchingFallbacks = ['release']
             debuggable false
