+++ app/proguard-benchmark.pro
@@ -1,0 +1,14 @@
+# Uncomment this to preserve the line number information for
+# debugging stack traces.
+-keepattributes SourceFile,LineNumberTable
+
+# If you keep the line number information, uncomment this to
+# hide the original source file name.
+-renamesourcefileattribute SourceFile
+
+# Repackage classes into the top-level.
+-repackageclasses
+
+# When generating the baseline profile we want the proper names of
+# the methods and classes
+-dontobfuscate
