+++ benchmark/src/main/java/com/numero/material_gallery/benchmark/BaselineProfileBenchmark.kt
@@ -1,0 +1,43 @@
+package com.numero.material_gallery.benchmark
+
+import androidx.benchmark.macro.BaselineProfileMode
+import androidx.benchmark.macro.CompilationMode
+import androidx.benchmark.macro.StartupMode
+import androidx.benchmark.macro.StartupTimingMetric
+import androidx.benchmark.macro.junit4.MacrobenchmarkRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class BaselineProfileBenchmark {
+    @get:Rule
+    val benchmarkRule = MacrobenchmarkRule()
+
+    @Test
+    fun startupNoCompilation() {
+        startup(CompilationMode.None())
+    }
+
+    @Test
+    fun startupBaselineProfile() {
+        startup(
+            CompilationMode.Partial(
+            baselineProfileMode = BaselineProfileMode.Require
+        ))
+    }
+
+    private fun startup(compilationMode: CompilationMode) {
+        benchmarkRule.measureRepeated(
+            packageName = "com.numero.material_gallery",
+            metrics = listOf(StartupTimingMetric()),
+            iterations = 10,
+            startupMode = StartupMode.COLD,
+            compilationMode = compilationMode
+        ) {
+            pressHome()
+            startActivityAndWait()
+        }
+    }
+}
